+++ src/containers/QueueContainer.js
@@ -1,25 +1,64 @@
-import React, {Component} from 'react';
-import {SwipeListView} from 'react-native-swipe-list-view';
+import React, { Component } from 'react';
+import { SwipeListView } from 'react-native-swipe-list-view';
 import isEmpty from 'lodash/isEmpty';
-import {Surface, Title, IconButton, Divider} from 'react-native-paper';
+import { Surface, Title, IconButton, Divider } from 'react-native-paper';
 import PropTypes from 'prop-types';
-import {connect} from 'react-redux';
-import { View, StyleSheet } from 'react-native';
+import { connect } from 'react-redux';
+import { View, StyleSheet, Alert } from 'react-native';
 
-import {clearQueue, removeFromQueue} from '../actions/playerState';
+import { clearQueue, removeFromQueue } from '../actions/playerState';
+import { getQueuedSongs } from '../actions/realmAction';
 import TrackContainer from './TrackContainer';
 import LoveContainer from './LoveContainer';
+import realm from '../database';
 
 class QueueContainer extends Component {
+  state = {
+    queue: []
+  }
   clearPlaylist = () => {
-    this.props.clearQueue();
+    Alert.alert(
+      'Clear Queue',
+      'Clear queue would stop current playing song',
+      [
+        {
+          text: 'Yes', onPress: () => {
+            this.props.close();
+            this.props.clearQueue();
+          }
+        },
+        {
+          text: 'Cancel',
+          onPress: () => console.log('Cancel Pressed'),
+          style: 'cancel',
+        },
+      ],
+      { cancelable: false },
+    );
+
   };
-  
+
+  componentDidMount() {
+    this.setState({
+      queue: getQueuedSongs()
+    })
+
+    realm.addListener('change', () => {
+      this.setState({
+        queue: getQueuedSongs()
+      })
+    })
+  }
+
+  componentWillUnmount() {
+    realm.removeAllListeners();
+  }
+
   render() {
-    return !isEmpty(this.props.queue) ? (
+    return !isEmpty(this.state.queue) ? (
       <View>
         <View style={styles.rowContainer}>
-          <Title style={{padding: 10}}>Queue</Title>
+          <Title style={{ padding: 10 }}>Queue</Title>
           <IconButton
             icon="delete"
             // size={40}
@@ -29,11 +68,11 @@ class QueueContainer extends Component {
 
         <Divider />
         <SwipeListView
-          data={this.props.queue}
-          renderItem={({item}) => <TrackContainer track={item} />}
+          data={this.state.queue}
+          renderItem={({ item }) => <TrackContainer track={item} />}
           ItemSeparatorComponent={() => <Divider inset={true} />}
           keyExtractor={(item, index) => index.toString()}
-          renderHiddenItem={({item}) => (
+          renderHiddenItem={({ item }) => (
             <Surface style={styles.rowBack}>
               <IconButton
                 icon="delete"
@@ -51,23 +90,20 @@ class QueueContainer extends Component {
         />
       </View>
     ) : (
-      false
-    );
+        false
+      );
   }
 }
 
-const mapStateToProps = state => ({
-  queue: state.playerState.queue,
-});
+
 
 QueueContainer.propTypes = {
-  queue: PropTypes.array.isRequired,
   clearQueue: PropTypes.func.isRequired,
   removeFromQueue: PropTypes.func.isRequired,
 };
 
 export default connect(
-  mapStateToProps,
+  null,
   {
     clearQueue,
     removeFromQueue,