+++ src/containers/TrackContainer.js
@@ -22,7 +22,7 @@ class TrackContainer extends React.Component {
   componentDidMount() {
     const {track, active} = this.props;
     if (!isUndefined(active)) {
-      if (isEqual(active.title, track.title)) {
+      if (isEqual(active.id, track.id)) {
         this.setState({
           isActive: true,
         });